scripts: gaia: add arg to autom install host deps & no cache
The --installHostDeps arg was added to the gaia script to allow the
automatic installation of host dependencies when building a recipe.
This need to be used with caution, as it will install and possible
mess up with host system. Is recommended to use it only when inside
a devcontainer;

The --noCache arg was added to the gaia script to allow the build
from scratch without any cache. This is behind the scene will
set the CLEAN_IMAGE env var to true, that is needed for firts
builds anyway.

Signed-off-by: Matheus Castello <matheus@castello.eng.br>